Fix broken arrows.
This was a pretty tricky one. It turns out that having arrows moving over
breakpoints comes with a number of issues. If we are changing the
arrow’s position from, say, top-left to bottom-right, we can’t just add
the new styles after a breakpoint; we actually have to undo the leftover
styles from the previous position (e.g. as well as adding `top: 100%;`,
we need to set `bottom: auto;`).

This got pretty hairy, as we also have to do the same unsetting for any
negative margins that pull the arrow around for middle positioning.

What I decided to do instead is to have breakpoint specific classes.
Now, instead of:

    <div class="right-box-arrow">

We’d have:

    <div class="bottom-box-arrow@sm  right-box-arrow@md">

Telling the `div` to only opt into the styles at separate breakpoints,
instead of undoing and overriding things.

Perhaps we should create a suite of Utility classes to abstract all of
this out…?