Fix broken arrows
This is a cleaner way of fixing the arrows (previous commit reverted).
We are now simply defensively resetting a bunch of stuff every time we
call a new arrow. This means there is a lot of verbosity and redundancy,
but it is the most self-contained and robust way of moving arrows around
responsively. Give me a shout if it doesn’t make sense.